Update optivum scrapper and crawler run script
