+++ packages/crawler/package.json
@@ -4,7 +4,7 @@
   "main": "dist/index.js",
   "scripts": {
     "build": "tsc",
-    "start": "node .",
+    "start": "node dist/run.js",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix"
   },
